refactored DeviceManager because it was  getting too big
