+++ DigitalPin.h
@@ -2,7 +2,7 @@
 #pragma once
 
 
-#include "DeviceManager.h"
+#include "NimbleAPI.h"
 
 class DigitalPin : public Device
 {
